[Tizen] Wait for mDNS on-browse events until timeout (#21553)
Tizen Native API dos not deliver all-for-now event (such event is
delivered by e.g. Avahi), so it is impossible to tell when the mDNS
browsing shall be considered finished. This commit adds timeout event
which is delivered after 250ms from the last on-browse event. This
allows Tizen platform to discover more than one mDNS service on local
network.